pyproject.toml
[tool.poetry]
name = "resarchpaper"
version = "0.1.0"
description = "CLI tool to fetch PubMed papers with non-academic authors"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "resarchpaper" }
]


[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.81"
typer = {extras = ["all"], version = "^0.12.3"}

[tool.poetry.scripts]
get-papers-list = "resarchpaper.cli:get_papers_list"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
